fix: remove meta info manager
